Tests/LibWeb/Text/expected/visual-viewport-horizontal-scroll-hit-test.txt

scale: 2
offsetLeft before scroll: 0
before scroll: left
offsetLeft after scroll: 400
after scroll: right
